:zap: Refactor all Client code
Removes a bunch of the abstractions and entrypoint interface modules.

Actually no longer needed, we just have the entrypoint spin up its own
process that handles IO, and we spin up a corresponding Client that sends
messages to it.

All generic IO hits the Client PID, which forwards it to the entrypoint's PID.