:bug: Dead entities should get removed from inventories
This bug happened when clients got disconnected, their Client process died,
and caused the linked Player entity process to also die.

This adds a message to be sent to all entities listening on an entity whenever
that target entity dies.

The Room uses an Inventory behaviour to keep contain all the entity pids, but
because the Room did not set up the observer correctly, it was not getting
the new killed event from the Player entity, leaving dead PIDs in the room.